+++ packages/orq-rc/src/mcp-server/mcp-server.ts
@@ -19,7 +19,7 @@ const routes = buildRouteMap({
 export const app = buildApplication(routes, {
   name: "mcp",
   versionInfo: {
-    currentVersion: "3.12.0-rc.18",
+    currentVersion: "3.12.0-rc.19",
   },
 });